Add mmio_dump command.
Implement a mmio_dump command that will print out the number of requested
bytes in 32-bit values. If the number of bytes requested is not a multiple of
4, it prints the remaining bytes as single bytes.

open_mapping() was modified to take a parameter of the desired bytes requested
to map. struct mmap_info was also modified to keep track of the correct length
of the mapping.