Add smbus_send_byte and smbus_receive_byte operations.
In order to properly support devices that do not know how to speak the SMBus
register access methods, the send byte and receive byte functionality needs
to be exported. An SMBUS_SIZE_BYTE operation instance was added to support this.
More checks were added to the default read/write paths to properly handle
changes in the number of arguments required for a given operation.