Add mem_readX and mem_dump commands
It's important to access the addressable memory space with the proper
type. Using the O_SYNC flag when opening /dev/mem signals to the kernel
to access the region as an uncacheable memory type. However, not using
O_SYNC tells the kernel to access the mapping/region as cacheable.
Therefore, add a new set of commands that provide the cacheable access.

Signed-off-by: Aaron Durbin <adurbin@gmail.com>